+++ sample-data/dictionaries.csv
@@ -1,0 +1,6 @@
+field2,field1,field3
+l1f2,l1f1,l1f3
+l2f2,l2f1,l2f3
+l3f2,l3f1,l3f3
+l4f2,l4f1,l4f3
+l5f2,l5f1,l5f3
