+++ tests/test_cli.py
@@ -1,0 +1,3 @@
+"""
+Unittests for newlinejson.__main__
+"""
